Add Mereldar's Toll direct DPS statistic
